Restructure directories to reflect install directory structure
