+++ lib/JitEngine.cpp
@@ -30,10 +30,10 @@
 #include <llvm/IR/Constants.h>
 #include <llvm/Transforms/Utils/ModuleUtils.h>
 
-#include "JitEngine.hpp"
+#include "proteus/JitEngine.hpp"
 
-#include "TimeTracing.hpp"
-#include "Utils.h"
+#include "proteus/TimeTracing.hpp"
+#include "proteus/Utils.h"
 
 // TODO: Used in InitTargetOptionsFromCodeGenFlags. Re-think for a different
 // initialization, especially using static compilation flags forwarded from
