+++ lib/proteus/JitEngineDevice.hpp
@@ -53,13 +53,13 @@
 #include <optional>
 #include <string>
 
-#include "CompilerInterfaceTypes.h"
-#include "JitCache.hpp"
-#include "JitEngine.hpp"
-#include "JitStorageCache.hpp"
-#include "TimeTracing.hpp"
-#include "TransformArgumentSpecialization.hpp"
-#include "Utils.h"
+#include "proteus/CompilerInterfaceTypes.h"
+#include "proteus/JitCache.hpp"
+#include "proteus/JitEngine.hpp"
+#include "proteus/JitStorageCache.hpp"
+#include "proteus/TimeTracing.hpp"
+#include "proteus/TransformArgumentSpecialization.hpp"
+#include "proteus/Utils.h"
 
 namespace proteus {
