+++ lib/proteus/JitVariable.hpp
@@ -7,7 +7,7 @@
 //===----------------------------------------------------------------------===//
 //
 //===----------------------------------------------------------------------===//
-#include "CompilerInterfaceTypes.h"
+#include "proteus/CompilerInterfaceTypes.h"
 
 #include <cstring>
