+++ pass/ProteusPass.cpp
@@ -62,11 +62,11 @@
 #include <llvm/Support/MemoryBuffer.h>
 #include <llvm/Support/MemoryBufferRef.h>
 
-#include "../common/Logger.hpp"
 #include <iostream>
 #include <string>
 
-#include "../lib/CompilerInterfaceTypes.h"
+#include "proteus/CompilerInterfaceTypes.h"
+#include "proteus/Logger.hpp"
 
 #include "GenRuntimeConstantTy.hpp"
